update auto terminate

+++ src/views/Create.js
@@ -38,17 +38,30 @@ function Create() {
   if(!sessionStorage.getItem('session')){
     window.location.href = '/login';}
 
+
+    
   var today = new Date();
   var time = today.getTime();
   var later=sessionStorage.getItem('time')-time;
+ 
+
+  if (performance.navigation.type == performance.navigation.TYPE_RELOAD){
+    console.log( "This page is reloaded" );
+    alert("Time Out");
+    sessionStorage.clear();
+    window.location.href = '/'
+  }
+
+  if(sessionStorage.getItem('flag')==1){
+  sessionStorage.setItem('flag', 0);
+    setTimeout(function(){
+    alert("Time Out");
+    sessionStorage.clear();
+    window.location.href = '/'
+  },later);
+  }
 
 
-  // alert(later);
-setTimeout(function(){
-  sessionStorage.clear();
-  alert("Time Out");
-  window.location.reload(false);
-},later);
 
 
   const [msisdn, setMsisdn] = useState(0);
@@ -2674,8 +2687,7 @@ setTimeout(function(){
   }
   
 
-
-window.addEventListener('load', function () {
+window.addEventListener('click', function () {
   const searchBar = document.getElementById('myInput');
 searchBar.addEventListener('keyup', (e) => {
       const searchString = e.target.value.toLowerCase();
