+++ src/views/Set.js
@@ -35,17 +35,13 @@ function Set() {
   if(!sessionStorage.getItem('session')){
     window.location.href = '/login';}
 
-  var today = new Date();
-  var time = today.getTime();
-  var later=sessionStorage.getItem('time')-time;
+    if (performance.navigation.type == performance.navigation.TYPE_RELOAD){
+      console.log( "This page is reloaded");
+      alert("Time Out");
+      sessionStorage.clear();
+      window.location.href = '/'
+    }
 
-
-  // alert(later);
-  setTimeout(function(){
-    sessionStorage.clear();
-  alert("Time Out");
-  window.location.reload(false);
-},later);
   
   
   // const [ serviceClass,sessionId,errordetails,showSob,sob,status,ip,username,imsi,profileID] = useState(0);
@@ -3854,7 +3850,7 @@ function Set() {
 
 
   
-window.addEventListener('load', function () {
+window.addEventListener('click', function () {
   const searchBar = document.getElementById('myInput');
 searchBar.addEventListener('keyup', (e) => {
       const searchString = e.target.value.toLowerCase();
